# packages/shared/openapi/openapi.generated.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Varasaan API",
    "version": "0.2.0",
    "description": "Contract-first backend API for Varasaan."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/v1/documents": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "listDocuments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Documents",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentSummaryResponse"
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/v1/documents/uploads/init": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "initDocumentUpload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload material issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "503": {
            "$ref": "#/components/responses/ServiceUnavailable"
          }
        }
      }
    },
    "/api/v1/documents/versions/{versionId}": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "getDocumentVersion",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/VersionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document version status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentVersionStatusResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/documents/versions/{versionId}/scan": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "queueDocumentScan",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/VersionId"
          }
        ],
        "responses": {
          "200": {
            "description": "Scan job enqueued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDispatchResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/documents/{documentId}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "operationId": "softDeleteDocument",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Soft deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "getDocument",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "responses": {
          "200": {
            "description": "Document detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDetailResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/documents/{documentId}/download": {
      "get": {
        "tags": [
          "documents"
        ],
        "operationId": "getDocumentDownloadUrl",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          },
          {
            "name": "trusted_contact_id",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Signed download URL",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDownloadResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/api/v1/documents/{documentId}/grants": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "createDocumentGrant",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GrantCreateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Grant created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{documentId}/versions/init": {
      "post": {
        "tags": [
          "documents"
        ],
        "operationId": "initDocumentVersionUpload",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/DocumentId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInitRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Upload material issued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadInitResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/v1/inventory/accounts": {
      "post": {
        "tags": [
          "inventory"
        ],
        "operationId": "createInventoryAccount",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Inventory account created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "inventory"
        ],
        "operationId": "listInventoryAccounts",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory accounts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/InventoryResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/inventory/accounts/{accountId}": {
      "put": {
        "tags": [
          "inventory"
        ],
        "operationId": "updateInventoryAccount",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/InventoryAccountId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InventoryUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inventory account updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "inventory"
        ],
        "operationId": "deleteInventoryAccount",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/InventoryAccountId"
          }
        ],
        "responses": {
          "200": {
            "description": "Inventory account deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiMessage"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "DocumentId": {
        "name": "documentId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
      "VersionId": {
        "name": "versionId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
      "InventoryAccountId": {
        "name": "accountId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication failed",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Access denied",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
...
      "ServiceUnavailable": {
        "description": "Upstream dependency unavailable",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiMessage": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "UploadInitRequest": {
        "type": "object",
        "required": [
          "doc_type",
          "size_bytes"
        ],
        "properties": {
          "doc_type": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer",
            "minimum": 1
          },
          "content_type": {
            "type": "string",
            "default": "application/octet-stream"
          },
          "sha256": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "UploadInitResponse": {
        "type": "object",
        "required": [
          "document_id",
          "version_id",
          "version_no",
          "object_key",
          "upload_url",
          "upload_url_expires_in_seconds",
          "plaintext_dek_b64",
          "kms_key_id"
        ],
        "properties": {
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "version_id": {
            "type": "string",
            "format": "uuid"
          },
          "version_no": {
            "type": "integer"
          },
          "object_key": {
            "type": "string"
          },
          "upload_url": {
            "type": "string"
          },
          "upload_url_expires_in_seconds": {
            "type": "integer"
          },
          "plaintext_dek_b64": {
            "type": "string"
          },
          "kms_key_id": {
            "type": "string"
          }
        }
      },
      "ScanDispatchResponse": {
        "type": "object",
        "required": [
          "version_id",
          "status"
        ],
        "properties": {
          "version_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "DocumentDownloadResponse": {
        "type": "object",
        "required": [
          "download_url",
          "expires_in_seconds"
        ],
        "properties": {
          "download_url": {
            "type": "string"
          },
          "expires_in_seconds": {
            "type": "integer"
          }
        }
      },
      "GrantCreateRequest": {
        "type": "object",
        "required": [
          "trusted_contact_id"
        ],
        "properties": {
          "trusted_contact_id": {
            "type": "string",
            "format": "uuid"
          },
          "granted_reason": {
            "type": "string",
            "nullable": true
          },
          "expires_in_hours": {
            "type": "integer",
            "nullable": true
          }
        }
      },
...
      "InventoryCreateRequest": {
        "type": "object",
        "required": [
          "platform",
          "category"
        ],
        "properties": {
          "platform": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "username_hint": {
            "type": "string",
            "nullable": true
          },
          "importance_level": {
            "type": "integer",
            "default": 2
          }
        }
      },
      "InventoryResponse": {
        "type": "object",
        "required": [
          "id",
          "platform",
          "category",
          "importance_level"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "platform": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "username_hint": {
            "type": "string",
            "nullable": true
          },
          "importance_level": {
            "type": "integer"
          }
        }
      },
...
      "InventoryUpdateRequest": {
        "type": "object",
        "required": [
          "platform",
          "category",
          "importance_level"
        ],
        "properties": {
          "platform": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "username_hint": {
            "type": "string",
            "nullable": true
          },
          "importance_level": {
            "type": "integer"
          }
        }
      },
      "DocumentVersionStatusResponse": {
        "type": "object",
        "required": [
          "id",
          "document_id",
          "version_no",
          "state",
          "object_key",
          "size_bytes",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "document_id": {
            "type": "string",
            "format": "uuid"
          },
          "version_no": {
            "type": "integer"
          },
          "state": {
            "type": "string"
          },
          "object_key": {
            "type": "string"
          },
          "size_bytes": {
            "type": "integer"
          },
          "sha256": {
            "type": "string",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "scan_status": {
            "type": "string",
            "nullable": true
          },
          "scan_summary": {
            "type": "string",
            "nullable": true
          },
          "scan_failed_purge_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "DocumentSummaryResponse": {
        "type": "object",
        "required": [
          "id",
          "doc_type",
          "state",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "doc_type": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "current_version_id": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "deleted_at": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "current_version": {
            "$ref": "#/components/schemas/DocumentVersionStatusResponse"
          }
        }
      },
      "DocumentDetailResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/DocumentSummaryResponse"
          },
          {
            "type": "object",
            "required": [
              "versions"
            ],
            "properties": {
              "versions": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/DocumentVersionStatusResponse"
                }
              }
            }
          }
        ]
      }
    }
  }
}